v0.6.0
## New commands:

"ask-me" - Gives a single awnser to the user, without creating an instance of the model requested.

"image" - Creates 4 images using DALL-E with a custom prompt.

"whisper" - Gives the user the ability to transcribe audio and also translate audio to english (only)

## Modified:

changeStatus - Create a check to see if the status is > 128 in length, using recursion to call the function again and redo it until it gives a valid status.

## New:

Moderation - Moderates the bot by using the OpenAi moderation tool.
systemInput  - To make my life easier.
handleSpecific - To handle specific commands in the bot's dms

## Other:

unsedCommands - Commands that are in WIP.
node_characterai - Files for custom typing since the repo doesnt support TS as of 27/04/2023.